+++ Source/ACE.Server/Command/Handlers/AdminCommands.cs
@@ -11,7 +11,6 @@ using ACE.Database;
 using ACE.Database.Models.Auth;
 using ACE.Database.Models.Shard;
 using ACE.Database.Models.World;
-using ACE.DatLoader;
 using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Entity.Enum.Properties;
@@ -3005,612 +3004,6 @@ namespace ACE.Server.Command.Handlers
                 Console.WriteLine("---------------");
             }
         }
-
-        [CommandHandler("verify-ranks", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Fixes various legacy bugs with attribute/vital/skill ranks and xp")]
-        public static void HandleVerifyRanks(Session session, params string[] parameters)
-        {
-            HandleVerifyAttributes(session, parameters);
-            HandleVerifyVitals(session, parameters);
-            HandleVerifySkills(session, parameters);
-        }
-
-        [CommandHandler("verify-skills", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Fixes various legacy bugs with skill ranks and xp")]
-        public static void HandleVerifySkills(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            uint totalSkillsVerified = 0;
-            uint totalPlayersVerified = 0;
-
-            var unknownSkillPlayers = new HashSet<uint>();
-
-            foreach (var player in players)
-            {
-                totalPlayersVerified++;
-                var updated = false;
-
-                foreach (var skill in player.Biota.BiotaPropertiesSkill)
-                {
-                    if (!Player.PlayerSkills.Contains((Skill)skill.Type))
-                    {
-                        unknownSkillPlayers.Add(player.Guid.Full);
-
-                        // if no pp/ranks, silently remove/continue?
-                        if (skill.PP == 0 && skill.LevelFromPP == 0)
-                            continue;
-
-                        Console.WriteLine($"{player.Name} has unknown skill {(Skill)skill.Type}");
-                        Console.WriteLine($"Advancement class: {(SkillAdvancementClass)skill.SAC}, Proficiency points: {skill.PP:N0}, Rank: {skill.LevelFromPP}");
-                        continue;
-                    }
-
-                    var rank = skill.LevelFromPP;
-
-                    var sac = (SkillAdvancementClass)skill.SAC;
-                    if (sac < SkillAdvancementClass.Trained)
-                        continue;
-
-                    totalSkillsVerified++;
-
-                    // verify skill rank
-                    var correctRank = Player.CalcSkillRank(sac, skill.PP);
-                    if (rank != correctRank)
-                    {
-                        Console.WriteLine($"{player.Name}'s {(Skill)skill.Type} rank is {rank}, should be {correctRank}");
-                        skill.LevelFromPP = (ushort)correctRank;
-                        updated = true;
-                    }
-
-                    // verify skill xp is within bounds
-
-                    // in retail, if a player had a trained skill maxed out, and then they speced it in spec temple,
-                    // they would sort of temporarily 'lose' that ~103m xp, unless they reset the trained skill, and then speced it
-
-                    // so the data can be in a legit situation here where a character has a skill speced,
-                    // but their xp is beyond the spec xp cap (4,100,490,438) and <= the trained xp cap (4,203,819,496)
-
-                    //var skillXPTable = Player.GetSkillXPTable(sac);
-                    var skillXPTable = Player.GetSkillXPTable(SkillAdvancementClass.Trained);
-                    var maxSkillXp = skillXPTable[skillXPTable.Count - 1];
-
-                    if (skill.PP > maxSkillXp)
-                    {
-                        var specTempleAte = skill.PP - maxSkillXp;
-                        Console.WriteLine($"{player.Name}'s {sac} {(Skill)skill.Type} skill total xp is {skill.PP:N0}, should be capped at {maxSkillXp:N0}");
-                        skill.PP = maxSkillXp;
-                        updated = true;
-
-                        // we could refund specTempleAte by adding it to AvailableExperience here
-                        // however, instead of operating with deltas, i think it would be much more consistent
-                        // to have a formula that adds up all of the XP consumed, and verify AvailableExperience = TotalExperience - calculated XP consumed
-                    }
-
-                }
-                //if (updated)
-                //player.SaveBiotaToDatabase();
-            }
-
-            Console.WriteLine($"Total players verified: {totalPlayersVerified:N0}");
-            Console.WriteLine($"Total skills verified: {totalSkillsVerified:N0}");
-            Console.WriteLine($"Unknown skill players: {unknownSkillPlayers.Count:N0}");
-        }
-
-        [CommandHandler("verify-heritage-augs", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Verifies all players have their heritage augs.")]
-        public static void HandleVerifyHeritageAugs(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            foreach (var player in players)
-            {
-                var heritage = (HeritageGroup?)player.GetProperty(PropertyInt.HeritageGroup);
-                if (heritage == null)
-                {
-                    Console.WriteLine($"Couldn't find heritage for {player.Name}");
-                    continue;
-                }
-
-                var heritageAug = GetHeritageAug(heritage.Value);
-                if (heritageAug == null)
-                {
-                    Console.WriteLine($"Couldn't find heritage aug for {heritage} player {player.Name}");
-                    continue;
-                }
-
-                var numAugs = player.GetProperty(heritageAug.Value) ?? 0;
-                if (numAugs < 1)
-                {
-                    Console.WriteLine($"{heritageAug}={numAugs} for {heritage} player {player.Name} -- fixing");
-
-                    player.SetProperty(heritageAug.Value, 1);
-                    player.SaveBiotaToDatabase();
-                }
-            }
-        }
-
-        public static Dictionary<PropertyInt, string> AugmentationDevices = new Dictionary<PropertyInt, string>()
-        {
-            { PropertyInt.AugmentationBonusImbueChance,             "gemaugmentationluckonimbues" },
-            { PropertyInt.AugmentationBonusSalvage,                 "gemaugmentationbonussalvage" },
-            { PropertyInt.AugmentationBonusXp,                      "gemaugmentationbonusxp" },
-            { PropertyInt.AugmentationCriticalDefense,              "gemaugmentationcriticaldefense" },
-            { PropertyInt.AugmentationCriticalExpertise,            "ace41482-eyeoftheremorseless" },
-            { PropertyInt.AugmentationCriticalPower,                "ace41481-handoftheremorseless" },
-            { PropertyInt.AugmentationDamageBonus,                  "ace41478-frenzyoftheslayer" },
-            { PropertyInt.AugmentationDamageReduction,              "ace41480-ironskinoftheinvincible" },
-            { PropertyInt.AugmentationExtraPackSlot,                "gemaugmentationpackslot" },
-            { PropertyInt.AugmentationFasterRegen,                  "gemaugmentationfastregen" },
-            { PropertyInt.AugmentationIncreasedCarryingCapacity,    "gemaugmentationcarryingcapacityi" },
-            { PropertyInt.AugmentationIncreasedSpellDuration,       "gemaugmentationspellduration" },
-            { PropertyInt.AugmentationInfusedCreatureMagic,         "ace41472-infusedcreaturemagic" },
-            { PropertyInt.AugmentationInfusedItemMagic,             "ace41473-infuseditemmagic" },
-            { PropertyInt.AugmentationInfusedLifeMagic,             "ace41474-infusedlifemagic" },
-            { PropertyInt.AugmentationInfusedVoidMagic,             "ace41479-infusedvoidmagic" },
-            { PropertyInt.AugmentationInfusedWarMagic,              "ace41475-infusedwarmagic" },
-            { PropertyInt.AugmentationInnateCoordination,           "gemaugmentationattcoordination" },
-            { PropertyInt.AugmentationInnateEndurance,              "gemaugmentationattendurance" },
-            { PropertyInt.AugmentationInnateFocus,                  "gemaugmentationattfocus" },
-            { PropertyInt.AugmentationInnateQuickness,              "gemaugmentationattquickness" },
-            { PropertyInt.AugmentationInnateSelf,                   "gemaugmentationattself" },
-            { PropertyInt.AugmentationInnateStrength,               "gemaugmentationattstrength" },
-            { PropertyInt.AugmentationJackOfAllTrades,              "ace43167-jackofalltrades" },
-            { PropertyInt.AugmentationLessDeathItemLoss,            "gemaugmentationdeathreduceditems" },
-            { PropertyInt.AugmentationResistanceAcid,               "gemaugmentationnaturalresistanceacid" },
-            { PropertyInt.AugmentationResistanceBlunt,              "gemaugmentationnaturalresistancebludg" },
-            { PropertyInt.AugmentationResistanceFire,               "gemaugmentationnaturalresistancefire" },
-            { PropertyInt.AugmentationResistanceFrost,              "gemaugmentationnaturalresistancefrost" },
-            { PropertyInt.AugmentationResistanceLightning,          "gemaugmentationnaturalresistanceelectric" },
-            { PropertyInt.AugmentationResistancePierce,             "gemaugmentationnaturalresistancepierc" },
-            { PropertyInt.AugmentationResistanceSlash,              "gemaugmentationnaturalresistanceslash" },
-            { PropertyInt.AugmentationSkilledMagic,                 "ace41476-masterofthefivefoldpath" },
-            { PropertyInt.AugmentationSkilledMelee,                 "ace41477-masterofthesteelcircle" },
-            { PropertyInt.AugmentationSkilledMissile,               "ace41490-masterofthefocusedeye" },
-            { PropertyInt.AugmentationSpecializeArmorTinkering,     "gemaugmentationtinkeringspecarmor" },
-            { PropertyInt.AugmentationSpecializeItemTinkering,      "gemaugmentationtinkeringspecitem" },
-            { PropertyInt.AugmentationSpecializeMagicItemTinkering, "gemaugmentationtinkeringspecmagic" },
-            { PropertyInt.AugmentationSpecializeSalvaging,          "gemaugmentationtinkeringspecsalv" },
-            { PropertyInt.AugmentationSpecializeWeaponTinkering,    "gemaugmentationtinkeringspecweap" },
-            { PropertyInt.AugmentationSpellsRemainPastDeath,        "gemaugmentationdeathspellsremain" },
-        };
-
-        [CommandHandler("verify-xp", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Verifies the amount of xp for all players")]
-        public static void HandleVerifyExperience(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            var diffPlayers = 0;
-            var results = new List<VerifyXpResult>();
-
-            foreach (var player in players)
-            {
-                var totalXP = player.GetProperty(PropertyInt64.TotalExperience) ?? 0;
-                var unassignedXP = player.GetProperty(PropertyInt64.AvailableExperience) ?? 0;
-
-                // loop through all attributes/vitals/skills, add up assigned xp
-                long attributeXP = 0;
-                long vitalXP = 0;
-                long skillXP = 0;
-                long augXP = 0;
-                long diffXP = Math.Min(0, player.GetProperty(PropertyInt64.VerifyXp) ?? 0);
-
-                foreach (var attribute in player.Biota.BiotaPropertiesAttribute)
-                    attributeXP += attribute.CPSpent;
-
-                foreach (var vital in player.Biota.BiotaPropertiesAttribute2nd)
-                    vitalXP += vital.CPSpent;
-
-                foreach (var skill in player.Biota.BiotaPropertiesSkill)
-                    skillXP += skill.PP;
-
-                var heritage = (HeritageGroup?)player.GetProperty(PropertyInt.HeritageGroup);
-                if (heritage == null)
-                    continue;       // ignore admins who have morphed into asheron / bael'zharon
-
-                var heritageAug = GetHeritageAug(heritage.Value);
-
-                foreach (var kvp in AugmentationDevices)
-                {
-                    var augProperty = kvp.Key;
-
-                    var numAugs = player.GetProperty(augProperty) ?? 0;
-                    if (augProperty == heritageAug)
-                        numAugs--;
-
-                    if (numAugs <= 0)
-                        continue;
-
-                    var aug = DatabaseManager.World.GetCachedWeenie(kvp.Value);
-                    var costPer = aug.GetProperty(PropertyInt64.AugmentationCost).Value;
-
-                    augXP += costPer * numAugs;
-                }
-
-                // todo: luminance auras
-
-                var calculatedSpent = attributeXP + vitalXP + skillXP + augXP + diffXP;
-
-                var currentSpent = totalXP - unassignedXP;
-
-                if (calculatedSpent != currentSpent)
-                {
-                    var diff = currentSpent - calculatedSpent;
-                    results.Add(new VerifyXpResult(player, calculatedSpent, currentSpent));
-                    diffPlayers++;
-                }
-            }
-
-            var xpList = DatManager.PortalDat.XpTable.CharacterLevelXPList;
-            var maxTotalXp = (long)xpList[xpList.Count - 1];
-
-            foreach (var result in results.OrderBy(i => i.Player.Name).OrderBy(i => i.Diff))
-            {
-                var player = result.Player;
-                var diff = result.Diff;
-
-                Console.WriteLine($"{player.Name}.HandleVerifyExperience() - calculatedSpent({result.Calculated:N0}) != currentSpent({result.Current:N0}) - diff({diff:N0}) -- fixed");
-
-                if (diff > 0)
-                {
-                    // add to unassigned xp
-                    var unassignedXP = player.GetProperty(PropertyInt64.AvailableExperience) ?? 0;
-                    player.SetProperty(PropertyInt64.AvailableExperience, unassignedXP + diff);
-                    player.SetProperty(PropertyInt64.VerifyXp, diff);
-                }
-                else
-                {
-                    var totalXP = player.GetProperty(PropertyInt64.TotalExperience) ?? 0;
-                    if (totalXP - diff > maxTotalXp)
-                    {
-                        Console.WriteLine($"{player.Name} has {totalXP:N0} total experience, {-diff:N0} would put them beyond max, attempting to subtract from unassigned..");
-                        var unassignedXP = player.GetProperty(PropertyInt64.AvailableExperience) ?? 0;
-                        if (unassignedXP + diff >= 0)
-                        {
-                            // do not assign diff in this case, or else re-running this command will still show differences for this branch
-                            player.SetProperty(PropertyInt64.AvailableExperience, unassignedXP + diff);
-                        }
-                        else
-                        {
-                            Console.WriteLine($"{player.Name} has {unassignedXP:N0} unassigned experience, {diff:N0} would put them < 0");
-                            continue;
-                        }
-                    }
-                    else
-                    {
-                        // setting the diff property below, which will be handled on player login
-                        // to properly handle possibly leveling up / skill credits
-                        player.SetProperty(PropertyInt64.VerifyXp, diff);
-                    }
-                }
-                player.SaveBiotaToDatabase();
-            }
-
-            Console.WriteLine($"\nDiff players: {diffPlayers:N0}");
-        }
-
-        public static PropertyInt? GetHeritageAug(HeritageGroup heritage)
-        {
-            switch (heritage)
-            {
-                case HeritageGroup.Aluvian:
-                case HeritageGroup.Gharundim:
-                case HeritageGroup.Sho:
-                case HeritageGroup.Viamontian:
-                    return PropertyInt.AugmentationJackOfAllTrades;
-
-                case HeritageGroup.Shadowbound:
-                case HeritageGroup.Penumbraen:
-                    return PropertyInt.AugmentationCriticalExpertise;
-
-                case HeritageGroup.Gearknight:
-                    return PropertyInt.AugmentationDamageReduction;
-
-                case HeritageGroup.Undead:
-                    return PropertyInt.AugmentationCriticalDefense;
-
-                case HeritageGroup.Empyrean:
-                    return PropertyInt.AugmentationInfusedLifeMagic;
-
-                case HeritageGroup.Tumerok:
-                    return PropertyInt.AugmentationCriticalPower;
-
-                case HeritageGroup.Lugian:
-                    return PropertyInt.AugmentationIncreasedCarryingCapacity;
-            }
-            return null;
-        }
-
-        [CommandHandler("verify-attributes", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Fixes various legacy bugs with attribute ranks and xp")]
-        public static void HandleVerifyAttributes(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            uint totalAttributesVerified = 0;
-            uint totalPlayersVerified = 0;
-
-            foreach (var player in players)
-            {
-                var updated = false;
-
-                foreach (var attr in player.Biota.BiotaPropertiesAttribute)
-                {
-                    var rank = attr.LevelFromCP;
-
-                    // verify attribute rank
-                    var correctRank = Player.CalcAttributeRank(attr.CPSpent);
-                    if (rank != correctRank)
-                    {
-                        Console.WriteLine($"{player.Name}'s {(PropertyAttribute)attr.Type} rank is {rank}, should be {correctRank}");
-                        attr.LevelFromCP = (ushort)correctRank;
-                        updated = true;
-                    }
-
-                    // verify attribute xp is within bounds
-                    var attributeXPTable = DatManager.PortalDat.XpTable.AttributeXpList;
-                    var maxAttributeXp = attributeXPTable[attributeXPTable.Count - 1];
-
-                    if (attr.CPSpent > maxAttributeXp)
-                    {
-                        Console.WriteLine($"{player.Name}'s {(PropertyAttribute)attr.Type} attribute total xp is {attr.CPSpent:N0}, should be capped at {maxAttributeXp:N0}");
-                        attr.CPSpent = maxAttributeXp;
-                        updated = true;
-                        // do AvailableExperience refund via non-delta method?
-                    }
-
-                    totalAttributesVerified++;
-                }
-
-                //if (updated)
-                    //player.SaveBiotaToDatabase();
-
-                totalPlayersVerified++;
-            }
-
-            Console.WriteLine($"Total players verified: {totalPlayersVerified:N0}");
-            Console.WriteLine($"Total attributes verified: {totalAttributesVerified:N0}");
-        }
-
-        [CommandHandler("verify-vitals", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Fixes various legacy bugs with vital ranks and xp")]
-        public static void HandleVerifyVitals(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            uint totalVitalsVerified = 0;
-            uint totalPlayersVerified = 0;
-
-            foreach (var player in players)
-            {
-                var updated = false;
-
-                foreach (var vital in player.Biota.BiotaPropertiesAttribute2nd)
-                {
-                    var rank = vital.LevelFromCP;
-
-                    // verify vital rank
-                    var correctRank = Player.CalcVitalRank(vital.CPSpent);
-                    if (rank != correctRank)
-                    {
-                        Console.WriteLine($"{player.Name}'s {(PropertyAttribute2nd)vital.Type} rank is {rank}, should be {correctRank}");
-                        vital.LevelFromCP = (ushort)correctRank;
-                        updated = true;
-                    }
-
-                    // verify vital xp is within bounds
-                    var vitalXPTable = DatManager.PortalDat.XpTable.VitalXpList;
-                    var maxVitalXp = vitalXPTable[vitalXPTable.Count - 1];
-
-                    if (vital.CPSpent > maxVitalXp)
-                    {
-                        Console.WriteLine($"{player.Name}'s {(PropertyAttribute2nd)vital.Type} vital total xp is {vital.CPSpent:N0}, should be capped at {maxVitalXp:N0}");
-                        vital.CPSpent = maxVitalXp;
-                        updated = true;
-                        // do AvailableExperience refund via non-delta method?
-                    }
-
-                    totalVitalsVerified++;
-                }
-
-                //if (updated)
-                    //player.SaveBiotaToDatabase();
-
-                totalPlayersVerified++;
-            }
-
-            Console.WriteLine($"Total players verified: {totalPlayersVerified:N0}");
-            Console.WriteLine($"Total vitals verified: {totalVitalsVerified:N0}");
-        }
-
-        [CommandHandler("verify-skill-credits", AccessLevel.Admin, CommandHandlerFlag.ConsoleInvoke, "Verifies and fixes player skill credits from Asheron's Castle.")]
-        public static void HandleVerifySkillCredits(Session session, params string[] parameters)
-        {
-            var players = PlayerManager.GetAllOffline();
-
-            using (var ctx = new ShardDbContext())
-            {
-                foreach (var player in players)
-                {
-                    // skip admins
-                    if (player.Account == null || player.Account.AccessLevel == (uint)AccessLevel.Admin)
-                        continue;
-
-                    // player starts with 52 skill credits
-                    var startCredits = 52;
-
-                    // skills that cannot be untrained: arcane lore, jump, loyalty, magic defense, run, salvaging
-                    // all of these have '0' cost to train, except for arcane lore, which has 4 (seems to be an outlier?)
-                    startCredits += 4;
-
-                    var levelCredits = GetAdditionalCredits(player.Level ?? 1);
-
-                    var totalCredits = startCredits + levelCredits;
-
-                    var used = 0;
-
-                    foreach (var skill in player.Biota.BiotaPropertiesSkill)
-                    {
-                        var sac = (SkillAdvancementClass)skill.SAC;
-                        if (sac < SkillAdvancementClass.Trained)
-                            continue;
-
-                        if (!DatManager.PortalDat.SkillTable.SkillBaseHash.TryGetValue(skill.Type, out var skillInfo))
-                        {
-                            Console.WriteLine($"{player.Name}.HandleVerifySkillCredits({(Skill)skill.Type}): unknown skill");
-                            continue;
-                        }
-
-                        //Console.WriteLine($"{(Skill)skill.Type} trained cost: {skillInfo.TrainedCost}, spec cost: {skillInfo.SpecializedCost}");
-
-                        used += skillInfo.TrainedCost;
-
-                        if (sac == SkillAdvancementClass.Specialized)
-                        {
-                            switch ((Skill)skill.Type)
-                            {
-                                // these can only be speced through augs, they have >= 999 in the spec data
-                                case Skill.ArmorTinkering:
-                                case Skill.ItemTinkering:
-                                case Skill.MagicItemTinkering:
-                                case Skill.WeaponTinkering:
-                                case Skill.Salvaging:
-                                    continue;
-                            }
-
-                            used += skillInfo.UpgradeCostFromTrainedToSpecialized;
-                        }
-                    }
-
-                    // 2 possible skill credits from quests
-                    // - ChasingOswaldDone
-                    // - ArantahKill1 (no 'turned in' stamp, only if given figurine?)
-                    var questCredits = ctx.CharacterPropertiesQuestRegistry.Count(i => i.CharacterId == player.Guid.Full && (i.QuestName.Equals("ChasingOswaldDone") || i.QuestName.Equals("ArantahKill1")));
-
-                    totalCredits += questCredits;
-
-                    // TODO: 2 lum augs
-
-                    var targetCredits = totalCredits - used;
-                    if (targetCredits < 0)
-                    {
-                        // if the player has already spent more skill credits than they should have,
-                        // unfortunately this situation requires a partial reset..
-                        Console.WriteLine($"{player.Name}.HandleVerifySkillCredits(): targetCredits({targetCredits}) < 0");
-                        UntrainSkills(player, targetCredits);
-                        continue;
-                    }
-
-                    var availableCredits = player.GetProperty(PropertyInt.AvailableSkillCredits) ?? 0;
-
-                    if (availableCredits != targetCredits)
-                    {
-                        Console.WriteLine($"{player.Name}.HandleVerifySkillCredits(): availableCredits({availableCredits}) != targetCredits({targetCredits}) -- fixing");
-                        player.SetProperty(PropertyInt.AvailableSkillCredits, targetCredits);
-                        player.SaveBiotaToDatabase();
-                    }
-
-                    //Console.WriteLine("--------------------");
-                }
-            }
-        }
-
-        private static void UntrainSkills(OfflinePlayer player, int targetCredits)
-        {
-            long refundXP = 0;
-
-            foreach (var skill in player.Biota.BiotaPropertiesSkill)
-            {
-                if (!DatManager.PortalDat.SkillTable.SkillBaseHash.TryGetValue(skill.Type, out var skillBase))
-                {
-                    Console.WriteLine($"{player.Name}.UntrainSkills({(Skill)skill.Type}) - unknown skill");
-                    continue;
-                }
-
-                var sac = (SkillAdvancementClass)skill.SAC;
-
-                if (sac != SkillAdvancementClass.Trained)
-                    continue;
-
-                refundXP += skill.PP;
-
-                skill.PP = 0;
-                skill.LevelFromPP = 0;
-
-                if (Player.IsSkillUntrainable((Skill)skill.Type))
-                {
-                    skill.SAC = (uint)SkillAdvancementClass.Untrained;
-                    skill.InitLevel -= 5;
-                    targetCredits += skillBase.TrainedCost;
-                }
-            }
-
-            var availableExperience = player.GetProperty(PropertyInt64.AvailableExperience) ?? 0;
-            player.SetProperty(PropertyInt64.AvailableExperience, availableExperience + refundXP);
-
-            player.SetProperty(PropertyInt.AvailableSkillCredits, targetCredits);
-
-            player.SetProperty(PropertyBool.UntrainedSkills, true);
-
-            player.SaveBiotaToDatabase();
-        }
-
-        public static int GetAdditionalCredits(int level)
-        {
-            foreach (var kvp in AdditionalCredits.Reverse())
-                if (level >= kvp.Key)
-                    return kvp.Value;
-
-            return 0;
-        }
-
-        /// <summary>
-        /// level => total additional credits
-        /// </summary>
-        public static SortedDictionary<int, int> AdditionalCredits = new SortedDictionary<int, int>()
-        {
-            { 2, 1 },
-            { 3, 2 },
-            { 4, 3 },
-            { 5, 4 },
-            { 6, 5 },
-            { 7, 6 },
-            { 8, 7 },
-            { 9, 8 },
-            { 10, 9 },
-            { 12, 10 },
-            { 14, 11 },
-            { 16, 12 },
-            { 18, 13 },
-            { 20, 14 },
-            { 23, 15 },
-            { 26, 16 },
-            { 29, 17 },
-            { 32, 18 },
-            { 35, 19 },
-            { 40, 20 },
-            { 45, 21 },
-            { 50, 22 },
-            { 55, 23 },
-            { 60, 24 },
-            { 65, 25 },
-            { 70, 26 },
-            { 75, 27 },
-            { 80, 28 },
-            { 85, 29 },
-            { 90, 30 },
-            { 95, 31 },
-            { 100, 32 },
-            { 105, 33 },
-            { 110, 34 },
-            { 115, 35 },
-            { 120, 36 },
-            { 125, 37 },
-            { 130, 38 },
-            { 140, 39 },
-            { 150, 40 },
-            { 160, 41 },
-            { 180, 42 },
-            { 200, 43 },
-            { 225, 44 },
-            { 250, 45 },
-            { 275, 46 }
-        };
 
         [CommandHandler("getenchantments", AccessLevel.Admin, CommandHandlerFlag.RequiresWorld, "Shows the enchantments for the last appraised item")]
         public static void HandleGetEnchantments(Session session, params string[] parameters)
